moved project to `phichain-chart`
